[huobi.pro] refactor ; delete old file
